IGNITE-24266 Move benchmark parameters to base class
